fix(product): correct catalog visibility, SKU edits, and review id width
- findAllProducts now defaults isActive to true. Without it the filter only
  applied when a caller happened to pass it, so the public storefront listed
  products their seller had hidden or the AI-02 risk scan had blocked. A
  single-seller scoped read keeps the old behaviour — that is the seller's
  own dashboard — and any caller can still ask for a state explicitly.
- The SKU tier validation now runs inside the update transaction. The SKU
  diff commits in its own transaction afterwards, so a bad skuList used to
  commit the new variation axes and then fail, leaving the product
  describing axes its SKUs no longer matched.
- The referenced-SKU lookup retries once. A stale TCP socket fails the first
  send and reconnects on the next, and that single blip silently downgraded
  a hard delete into a deactivation, so a removed variant lingered as an
  inactive row behind a 200.
- sku_upserted now carries stockChangedSkuIds so inventory can tell a real
  restock from an echo of the product mirror.
- product_reviews.product_id maps as bigint to match products.id (migration
  nodeA-20260804-001). TypeORM maps bigint to a JS string, and the gateway
  already rewrites productId to the prod_ public id on both the create and
  the list path, so the HTTP contract is unchanged.

+++ apps/product/src/entity/product-review.entity.ts
@@ -13,7 +13,9 @@ export class ProductReview {
   @PrimaryGeneratedColumn()
   id!: number;
 
-  @Column({ name: "product_id", type: "int" })
+  // bigint to match products.id and the other FKs pointing at it
+  // (product_skus, wishlist_items, product_risk_feedback).
+  @Column({ name: "product_id", type: "bigint" })
   productId!: number;
 
   @Column({ name: "user_id", type: "int" })
